Cleanup some leftovers from refactoring
