+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/CKEditor5/plugins/ExternalLinkPlugin/constants.js
@@ -1,0 +1,5 @@
+// @flow
+const LINK_HREF_ATTRIBUTE = 'linkHref';
+const LINK_TARGET_ATTRIBUTE = 'linkTarget';
+
+export {LINK_HREF_ATTRIBUTE, LINK_TARGET_ATTRIBUTE};
